Simplify dagger logic
