+++ plugin/dagger/src/main/java/io/deephaven/plugin/PluginModule.java
@@ -8,10 +8,18 @@ import io.deephaven.plugin.type.ObjectTypeModule;
 
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.ServiceLoader;
+import java.util.ServiceLoader.Provider;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
- * Provides an empty set for {@link Plugin} and adapters into {@link Plugin} from {@link ObjectType}.
+ * Provides the set of {@link Registration} from {@link ServiceLoader#load(Class)} and as adapted from {@link Plugin}
+ * via {@link PluginAdapter}.
+ *
+ * <p>
+ * Provides the set of {@link Plugin} from {@link ServiceLoader#load(Class)} and as casted directly from
+ * {@link ObjectType}.
  *
  * @see ObjectTypeModule
  */
@@ -20,8 +28,20 @@ public interface PluginModule {
 
     @Provides
     @ElementsIntoSet
-    static Set<Plugin> primesPlugins() {
-        return Collections.emptySet();
+    static Set<Registration> providesServiceLoaderRegistrations() {
+        return ServiceLoader.load(Registration.class).stream().map(Provider::get).collect(Collectors.toSet());
+    }
+
+    @Provides
+    @ElementsIntoSet
+    static Set<Registration> adaptsPlugins(Set<Plugin> plugins) {
+        return plugins.stream().map(PluginAdapter::new).collect(Collectors.toSet());
+    }
+
+    @Provides
+    @ElementsIntoSet
+    static Set<Plugin> providesServiceLoaderPlugins() {
+        return ServiceLoader.load(Plugin.class).stream().map(Provider::get).collect(Collectors.toSet());
     }
 
     @Provides
