+++ plugin/dagger/src/main/java/io/deephaven/plugin/type/ObjectTypeModule.java
@@ -3,19 +3,22 @@ package io.deephaven.plugin.type;
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.ElementsIntoSet;
+import io.deephaven.plugin.Plugin;
 
-import java.util.Collections;
+import java.util.ServiceLoader;
+import java.util.ServiceLoader.Provider;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
- * Provides an empty set for {@link ObjectType}.
+ * Provides the set of {@link ObjectType} from {@link ServiceLoader#load(Class)}.
  */
 @Module
 public interface ObjectTypeModule {
 
     @Provides
     @ElementsIntoSet
-    static Set<ObjectType> primesObjectTypes() {
-        return Collections.emptySet();
+    static Set<ObjectType> providesServiceLoaderObjectTypes() {
+        return ServiceLoader.load(ObjectType.class).stream().map(Provider::get).collect(Collectors.toSet());
     }
 }
